Undocumented ENV variable: AWS_S3_BUCKET_FAVICONS
I am not sure with there is a different variable in the first place, but after spending a lot of time debugging my favicons, my understanding is that it's required.
This commit documents it in the example env file and add a fallback to AWS_S3_BUCKET if it's not set.